ai-service/requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
groq
boto3
python-dotenv
firebase-admin
python-multipart
requests
